Create pymongo abstraction for our application
Add classes User, Listing, and Database

User represents a User that will be added to the database. Note that the password should already be hashed prior to being added to this object.

Listing represents a Listing that will be added to the database. This will likely change as we further grasp the scope of the data we want to track for each listing.

Database represents the mongodb database we are using for this project and provides a simple interface to the database and adds a layer of abstraction for simplicity and ease of later edits.