Connected REACT to Flask to get listing data
I created a new function get_listenings to pull all the listing data from the database. Then in index.jsx, I added a way for REACT to get that data as well. This will help create a list of listings in the future to help the person implementing REACT with listing data in the future.